feat: 채팅 float icon button으로 수정
